# app/static/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "GT20 Team Final Project of Backend Track B-)",
        "version": "0.0.1",
        "title": "Startup Campus Backend Final Project",
        "contact": {
            "name": "Team GT20",
            "email": "gt20@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://fashion-campus-gt20.my.id:5000"
        }
    ],
    "components": {
        "securitySchemes": {
            "Fashion-Campus-Auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header",
                "name": "JWT-FC"
            }
        },
...
    },
    "paths": {
...
        "/order": {
            "post": {
                "tags": [
                    "CART"
                ],
                "summary": "Process an order from user",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "properties": {
                                    "shipping_method": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    }
                                }
                            },
                            "examples": {
                                "success": {
                                    "description": "Success created an order",
                                    "value": [
                                        {
                                            "shipping_method": "Same day"
                                        },
                                        {
                                            "name": "Anthony",
                                            "phone_number": "082145427945",
                                            "address": "Jl Manchester City Etihad Stadium",
                                            "city": "Manchester"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Order success",
                                        "value": {
                                            "message": "Order success"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "tags": [
                    "PROFILE PAGE"
                ],
                "summary": "Get user's profile in detail",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success get user's profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "phone_number": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Get user's profile success",
                                        "value": {
                                            "name": "Anthony",
                                            "email": "anthony@manchester.com",
                                            "phone_number": "082145427945"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
        "/user/balance": {
            "post": {
                "tags": [
                    "PROFILE PAGE"
                ],
                "summary": "Add balance to user's money",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string",
                                "properties": {
                                    "amount": {
                                        "type": "integer"
                                    }
                                }
                            },
                            "examples": {
                                "success": {
                                    "description": "Success create balance",
                                    "value": {
                                        "amount": 1000000
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Success create balance",
                                        "value": {
                                            "message": "Top up balance success"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            },
            "get": {
                "tags": [
                    "PROFILE PAGE"
                ],
                "summary": "Get current user's balance",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success getting user's balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "balance": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Success getting user's balance",
                                        "value": {
                                            "balance": 1000000
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
        "/user/order": {
            "get": {
                "tags": [
                    "PROFILE PAGE"
                ],
                "summary": "Get all information from user's order",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success get user's order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "details": {
                                                        "type": "object",
                                                        "properties": {
                                                            "quantity": {
                                                                "type": "integer"
                                                            },
                                                            "size": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "price": {
                                                        "type": "integer"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "shipping_method": {
                                            "type": "string"
                                        },
                                        "shipping_address": {
                                            "type": "object",
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "type": "string"
                                                },
                                                "city": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Success get user's order detail",
                                        "value": [
                                            {
                                                "id": "bd93ccda-2342-4698-aefc-fc44e187be26",
                                                "created_at": "Fri, 27 August 2022",
                                                "product": [
                                                    {
                                                        "id": "356ffd58-72d1-40f9-9e39-8bdf3020ccaa",
                                                        "details": {
                                                            "quantity": 20,
                                                            "size": "L"
                                                        },
                                                        "price": 100000,
                                                        "image": "/static/blazer_barcelona.png",
                                                        "name": "Decul Garis Keras"
                                                    }
                                                ],
                                                "shipping_method": "Same day",
                                                "shipping_address": {
                                                    "address_name": "Anthony Milinkovic",
                                                    "phone_number": "082145421945",
                                                    "address": "Jl Juventus Turin Allianz Stadium",
                                                    "city": "Turin"
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
...
    }
}